requirements.txt
flask
google-generativeai
faiss-cpu
sentence-transformers
numpy